requirements.txt
sseclient-py==1.7
requests==2.24.0
numpy==1.19.1
pandas==1.1.0
matplotlib==3.3.0
